requirements.txt
getkey==0.6.5
pytest==8.3.4
pytest-mock==3.14.0
pyway==0.3.32 
python-dotenv==1.0.1
psycopg2-binary==2.9.10
polygon-api-client